backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9
bcrypt==4.2.0
PyJWT==2.9.0
email-validator==2.2.0
